Remove castings

+++ src/interfaces/IUserProxy.sol
@@ -50,5 +50,5 @@ interface IUserProxy {
     function unstake(uint256 _amount, address _recipient) external returns (uint256 lusdAmount, uint256 ethAmount);
     /// @notice Returns the current amount LQTY staked by a user in the V1 staking contract
     /// @return staked Amount of LQTY tokens staked
-    function staked() external view returns (uint88);
+    function staked() external view returns (uint256);
 }
