+++ test/UniV4Donations.t.sol
@@ -64,8 +64,8 @@ abstract contract UniV4DonationsTest is Test, Deployers {
     uint128 private constant REGISTRATION_FEE = 1e18;
     uint128 private constant REGISTRATION_THRESHOLD_FACTOR = 0.01e18;
     uint128 private constant UNREGISTRATION_THRESHOLD_FACTOR = 4e18;
-    uint16 private constant REGISTRATION_WARM_UP_PERIOD = 4;
-    uint16 private constant UNREGISTRATION_AFTER_EPOCHS = 4;
+    uint256 private constant REGISTRATION_WARM_UP_PERIOD = 4;
+    uint256 private constant UNREGISTRATION_AFTER_EPOCHS = 4;
     uint128 private constant VOTING_THRESHOLD_FACTOR = 0.04e18;
     uint88 private constant MIN_CLAIM = 500e18;
     uint88 private constant MIN_ACCRUAL = 1000e18;
@@ -147,7 +147,7 @@ abstract contract UniV4DonationsTest is Test, Deployers {
 
         vm.startPrank(lusdHolder);
         assertEq(uniV4Donations.donateToPool(), 0, "d");
-        (uint240 amount, uint16 epoch, uint256 released) = uniV4Donations.vesting();
+        (uint256 amount, uint256 epoch, uint256 released) = uniV4Donations.vesting();
         assertEq(amount, 1000e18, "amt");
         assertEq(epoch, 1, "epoch");
         assertEq(released, 0, "released");
@@ -198,7 +198,7 @@ abstract contract UniV4DonationsTest is Test, Deployers {
 
         vm.startPrank(lusdHolder);
         assertEq(uniV4Donations.donateToPool(), 0, "d");
-        (uint240 amount, uint16 epoch, uint256 released) = uniV4Donations.vesting();
+        (uint256 amount, uint256 epoch, uint256 released) = uniV4Donations.vesting();
         assertEq(amount, amt, "amt");
         assertEq(epoch, 1, "epoch");
         assertEq(released, 0, "released");
