+++ test/recon/properties/TsProperties.sol
@@ -11,7 +11,7 @@ abstract contract TsProperties is BeforeAfter {
 
     function property_user_ts_is_always_greater_than_start() public {
         for (uint256 i; i < users.length; i++) {
-            (uint88 user_allocatedLQTY, uint120 userTs) = governance.userStates(users[i]);
+            (uint256 user_allocatedLQTY, uint256 userTs) = governance.userStates(users[i]);
             if (user_allocatedLQTY > 0) {
                 gte(userTs, magnifiedStartTS, "User ts must always be GTE than start");
             }
@@ -19,7 +19,7 @@ abstract contract TsProperties is BeforeAfter {
     }
 
     function property_global_ts_is_always_greater_than_start() public {
-        (uint88 totalCountedLQTY, uint120 globalTs) = governance.globalState();
+        (uint256 totalCountedLQTY, uint256 globalTs) = governance.globalState();
 
         if (totalCountedLQTY > 0) {
             gte(globalTs, magnifiedStartTS, "Global ts must always be GTE than start");
